requirements.txt
flask
google-generativeai
sentence-transformers
numpy
faiss-cpu
pypdf
flask-cors
requests
mysql-connector-python